Allow prometheus exporter namespace to be configurable.
Currently the prometheus exporter namespace is empty, which means
metrics are rooted at 'opencensus_io'. This is less than ideal
and currently not configurable. This change proposes adding a new
prometheus exporter configuration param which allows the setting of
a custom namespace. In the event this parameter is not set, the
namespace will default to 'krakend'.